# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "E-Commerce Core API Documentation",
        "description": "API documentation for Laravel E-Commerce Core project including Auth, Products, Categories, Cart, Orders, Coupons, Reviews/Comments, Addresses, and Admin Management.",
        "contact": {
            "name": "API Support Team",
            "email": "admin@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Danh sách sản phẩm (Có tìm kiếm & lọc)",
                "operationId": "d97d1c3727cbac49708b85f7d11f7d58",
                "parameters": [
                    {
                        "name": "keyword",
                        "in": "query",
                        "description": "Từ khóa tìm kiếm theo tên/mô tả",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "category_id",
                        "in": "query",
                        "description": "ID danh mục",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "min_price",
                        "in": "query",
                        "description": "Giá tối thiểu",
                        "required": false,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "max_price",
                        "in": "query",
                        "description": "Giá tối đa",
                        "required": false,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "sort_by",
                        "in": "query",
                        "description": "Trường sắp xếp (price, created_at, name)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "created_at"
                        }
                    },
                    {
                        "name": "sort_order",
                        "in": "query",
                        "description": "Hướng sắp xếp (asc, desc)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "desc"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Số phần tử trên 1 trang",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Danh sách sản phẩm phân trang"
                    }
                }
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Tạo mới sản phẩm (Admin/Staff)",
                "operationId": "785644b7d7dd38af4b51dafd0fbc1b6d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "category_id",
                                    "name",
                                    "sku",
                                    "price",
                                    "stock",
                                    "description",
                                    "status"
                                ],
                                "properties": {
                                    "category_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Áo Polo Nam Premium Cotton"
                                    },
                                    "sku": {
                                        "type": "string",
                                        "example": "POLO-NAM-99"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 350000
                                    },
                                    "sale_price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 299000
                                    },
                                    "stock": {
                                        "type": "integer",
                                        "example": 50
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Áo polo chất liệu cotton co dãn 4 chiều, thoáng mát."
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "published"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tạo sản phẩm thành công"
                    },
                    "403": {
                        "description": "Không có quyền thực hiện"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/products/{product}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Chi tiết sản phẩm",
                "operationId": "db00b523e69d5dbc09c6edad76374bdc",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "ID hoặc slug sản phẩm",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "iphone-15-pro-max-256gb"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Chi tiết sản phẩm"
                    },
                    "404": {
                        "description": "Sản phẩm không tồn tại"
                    }
                }
            },
            "put": {
                "tags": [
                    "Products"
                ],
                "summary": "Cập nhật sản phẩm (Admin/Staff)",
                "operationId": "f47cf55f4f3e0b134b5e7d9a8373725d",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "ID sản phẩm",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Sản phẩm đã sửa"
                                    },
                                    "price": {
                                        "type": "number",
                                        "example": 199000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cập nhật thành công"
                    },
                    "403": {
                        "description": "Không có quyền"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Products"
                ],
                "summary": "Xóa sản phẩm (Admin/Staff)",
                "operationId": "ac878db81fdb1333b21fc79e574a5a17",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "ID sản phẩm",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Xóa thành công"
                    },
                    "403": {
                        "description": "Không có quyền"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "sanctum": {
                "type": "http",
                "description": "Enter token in format: Bearer {your_token}",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}